Struct nic_buffers::ReceiveBuffer

pub struct ReceiveBuffer {
    pub mp: MappedPages,
    pub phys_addr: PhysicalAddress,
    pub length: u16,
    /* private fields */
}

A buffer that stores a packet (a piece of an Ethernet frame) that has been received from the NIC and is guaranteed to be contiguous in physical memory. Auto-dereferences into a MappedPages object that represents its underlying memory. When dropped, its underlying memory is automatically returned to the NIC driver for future reuse.

Fields
mp: MappedPagesphys_addr: PhysicalAddresslength: u16
Implementations

impl ReceiveBuffer

pub fn new(
mp: MappedPages,
phys_addr: PhysicalAddress,
length: u16,
pool: &'static Queue<ReceiveBuffer>
) -> ReceiveBuffer

Creates a new ReceiveBuffer with the given MappedPages, PhysicalAddress, and length. When this ReceiveBuffer object is dropped, it will be returned to the given pool.